Add mouseover functionality.
Cells are now highlighted when you mouseover them and their time and charge
is shown in the status bar.  This only works when not animating or free running.

Also, do not crash if the file has zero events.  There's not much useful
can be done in this case, but at least show a message saying what the
problem is.