Buffer 25 events up front.
Instead of starting with 1 event, start with 25, or the whole file
if there are less than 25.  This pushes away some trouble I have with
receiving X events while fetching detector events that can make the
GUI unresponsive.  It would be a better solution to fix this logic,
although I'm not sure it can be entirely fixed in one thread with
art as is.