Continue animation as expected.
Fix the case of changing the speed while animating and free running.
It now continues animating.

Also fix a minor bug where we set the mouseover timer handler to
-1, except it is unsigned.

Close loophole in draw_event.  The animating flag is set before we
check for X events so the above speed-change fix should always work.