Fix prefetching+free running.
Previously, if prefetching and free running at speed 11, the
free run interval would be zero, and we'd get locked into an
infinite loop of trying to get the next event, but not getting
it.  Now limit the interval to be at least 1ms.

Also avoid calling g_main_context_iteration if in the middle of
a prefetch.

Oh, and it turns out that I failed to ever clear the "prefetching"
flag.  That is also fixed.