Add spin button for selecting tick.
Only the max tick is selectable.  I'd also like the min tick so the
user can select a time window.

The design is getting obviously unwieldly.  Should refactor so that
there's a function that draws one animation frame that's called with
a timer, and then lots of these globals and fragile tests would go
away.