Refactor to use GTK more properly.
Now handle all timing with g_timeout and not with usleep
and calls to gtk_context_iteration or whatever. This removes lots
of fiddling that was necessary to make the old fiddly way work.

The current functionality is the same as before except that for
the moment animation and free running are mutually exclusive
because if you turn them both on, their timers run asynchronously
and don't do what anyone would want.