Refactor notions of time ranges.
Now recognizes explicitly that there are four relevant time ranges:
the whole event, the part the user has selected, the part that is
visible now, and the part that is to be drawn now.  Much less
confusing now that I know it is confusing.

Noted a variety of deficiencies with TODOs.