Draws tracks by trajectory.
Somewhat sensible drawing of tracks by trajectory points
rather than by the set of hits.  Still hit based, so if you
zoom in it is not quite the correct trajectory.  Also if you zoom
in you lose any track segment if any part of it is outside the
drawing area, which isn't very desireable.