Don't draw track segments out of range.
Check both ends of a track segment and don't draw it if it goes
outside the view.  This is correct, but not very good.  Much
better would be to separate the drawing area into two drawing
areas, one for each view, and draw the segment if *either* end
were inside, letting Cairo handle the edges.