Find number of events up front.
Read in the number of events in the file up front and
use this to display the fraction loaded so far instead
of just a running total.  Make the status line that
one gets if one enters and invalid event nicer, particularly
in that it doesn't dissappear based on an old timer
anymore.