Find screen track points once.
Instead of finding the screen positions of track points
twice for every point except the first and last by finding
them within draw_trackseg, find them all first and then
pass the results in.

This is not done so much for efficiency as so that the
positions can later be cached and used for mouseover features.