Add track highlighting.
It is now in a reasonable shape visually.  Performance isn't great
for events with lots of hits since it redraws the whole event when
the highlighting changes.  No information is shown about the
highlighted track.  I think I'd like that in a new window, 90's
Unix style.  But at least it matches the views for you.