Improve track highlighting performance.
Save cairo pattern of all the hits when we draw so that when we
change the highlighting of tracks, we can restore it instead of
having to draw them all again.

I hope that the saving itself isn't expensive, since if we don't
have tracks or aren't mousing over them, this doesn't buy anything.