Add new status line for tracks.
Now the 3rd status line shows track information while the 4th shows
error messages.

There are getting to be a lot of status lines.  It would be better to have
a pop-up showing track information that could show, e.g., all the trajectory
points.  (Or a side pane or whatever turns out to be easy to implement
in GTK+.)  The same goes for the hit information, for that matter.

Currently the only information shown is the track index and the start
and stop points, with mm precision.