Draw active track on top.
Always draw the active track last so that it appears on top
of the other tracks.  This is particularly important with
BreakPointFitter, which always gives you three nearly-identical
tracks almost on top of each other.  If the active one is
on the bottom, you sometimes can't see that it's selected at all.