Poll mouse position.
Poll mouse position twice a second so that the user need not
wiggle the mouse to get information on hits and tracks when
the event has changed or the animation has progressed.  Also
helps get the right hit/track highlighted when it is a busy
event and the user has swept the mouse over lots of objects.
It seems in this case, and perhaps I am misinterpreting, that
GTK "hinting" doesn't always do a good job, and so no
motion-notify-event is generated for the final position.