Protect against partially misssing calhits.
Apparently it is possible for some events to have a product
named "calhit" but for others to lack it.  This is the case
for fardet_r00022563_s47_t02_R17-03-01-prod3reco.h_v1_data.pid.root

Chris Backhouse speculates that this is because removebeamspills
is run before calhit.  I would have thought you'd just get empty
calhits in that case, but maybe not.

This solution is non-ideal because the feedback to the user is
only in the shell and not the GUI.  With the old assumption that
a file either has calhits or not, this was ok because the GUI just
wouldn't start up at all if they were missing.  Maybe it's still
sufficient because the events missing calhits aren't wanted anyway.